# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tada",
    "description": "The tada HTTP API. The list of problem codes of each operation is open.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1"
  },
  "paths": {
    "/api/v1/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Lists the events that the caller can see, in the order of their keys.",
        "operationId": "list_events",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "The page size: 1 to 200. The default is 50.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 200,
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "The `next_cursor` of the previous page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of events.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            }
          },
          "default": {
            "description": "A problem (ADR 0037).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-tada-problem-codes": [
          "malformed-request",
          "unauthenticated",
          "organization-required",
          "unavailable",
          "internal"
        ]
      },
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Creates an event in the caller's organization.",
        "description": "If an event with the same `id` and the same content exists, the response is this event with the\nstatus 200, and nothing changes.",
        "operationId": "create_event",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The event exists with this ID and the same content.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "201": {
            "description": "The new event.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "default": {
            "description": "A problem (ADR 0037).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-tada-problem-codes": [
          "malformed-request",
          "unauthenticated",
          "organization-required",
          "forbidden",
          "payload-too-large",
          "unsupported-media-type",
          "validation-failed",
          "unavailable",
          "internal"
        ]
      }
    },
    "/api/v1/events/{event_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Reads one event. The caller needs an event role in it, or the organization role owner or admin.",
        "operationId": "get_event",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "description": "The ID of the event.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The event.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "default": {
            "description": "A problem (ADR 0037).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-tada-problem-codes": [
          "malformed-request",
          "unauthenticated",
          "organization-required",
          "not-found",
          "unavailable",
          "internal"
        ]
      }
    },
...
    "/api/v1/sign-in/magic-link": {
      "post": {
        "tags": [
          "sign-in"
        ],
        "summary": "Signs in with the token of a magic link. The token works once.\nThe response sets the session cookie.",
        "operationId": "redeem_magic_link",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemMagicLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The new session. The `Set-Cookie` header holds its token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            }
          },
          "default": {
            "description": "A problem (ADR 0037).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-tada-problem-codes": [
          "malformed-request",
          "unauthenticated",
          "forbidden",
          "payload-too-large",
          "unsupported-media-type",
          "unavailable",
          "internal"
        ]
      }
    },
...
    "/api/v1/invitations/preview": {
      "post": {
        "tags": [
          "sign-in"
        ],
        "summary": "Shows the organization and the role of an invitation. It does not use the token.",
        "operationId": "preview_invitation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "What the invitation is for.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationPreview"
                }
              }
            }
          },
          "default": {
            "description": "A problem (ADR 0037).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-tada-problem-codes": [
          "malformed-request",
          "unauthenticated",
          "forbidden",
          "payload-too-large",
          "unsupported-media-type",
          "unavailable",
          "internal"
        ]
      }
    },
    "/api/v1/invitations/accept": {
      "post": {
        "tags": [
          "sign-in"
        ],
        "summary": "Accepts an invitation. The token works once. The response sets the cookie of a new session\nin the organization of the invitation.",
        "operationId": "accept_invitation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The new session. The `Set-Cookie` header holds its token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            }
          },
          "default": {
            "description": "A problem (ADR 0037).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-tada-problem-codes": [
          "malformed-request",
          "unauthenticated",
          "forbidden",
          "payload-too-large",
          "unsupported-media-type",
          "unavailable",
          "internal"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventPage": {
        "type": "object",
        "description": "One page of events.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "The cursor of the next page. It is absent on the last page."
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "An error response. `detail` never repeats input values.",
        "required": [
          "type",
          "code",
          "title",
          "status",
          "instance",
          "request_id"
        ],
        "properties": {
          "type": {
            "type": "string",
            "description": "The URL of the code in the public catalog."
          },
          "code": {
            "type": "string",
            "description": "The stable problem code. Clients react to this field. The list of codes is open."
          },
          "title": {
            "type": "string",
            "description": "A short, stable English text for developers."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "An English text about this case, for developers. Clients never show it to members."
          },
          "instance": {
            "type": "string",
            "description": "`urn:uuid:` and the request ID."
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemError"
            },
            "description": "The invalid values of a `validation-failed` problem."
          }
        }
      },
...
      "RedeemMagicLinkRequest": {
        "type": "object",
        "description": "The input of a sign-in with a magic link.",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string",
            "description": "The token from the fragment of the magic link."
          }
        }
      },
...
    }
  }
}